+++ packages/frontend/i18n/src/resources/en.json
@@ -811,6 +811,7 @@
   "com.affine.payment.ai.benefit.g3-3": "Open source & Privacy ensured",
   "com.affine.payment.ai.billing-tip.end-at": "You have purchased AFFiNE AI. The expiration date is {{end}}.",
   "com.affine.payment.ai.billing-tip.next-bill-at": "You have purchased AFFiNE AI. The next payment date is {{due}}.",
+  "com.affine.payment.billing-tip.past-due": "Your recent payment failed, the next payment date is {{due}}.",
   "com.affine.payment.ai.pricing-plan.caption-free": "You are currently on the Free plan.",
   "com.affine.payment.ai.pricing-plan.caption-purchased": "You have purchased AFFiNE AI",
   "com.affine.payment.ai.pricing-plan.learn": "Learn about AFFiNE AI",
@@ -824,6 +825,9 @@
   "com.affine.payment.ai.usage.purchase-button-label": "Purchase",
   "com.affine.payment.ai.usage.used-caption": "Times used",
   "com.affine.payment.ai.usage.used-detail": "{{used}}/{{limit}} times",
+  "com.affine.payment.subscription-status.active": "Active",
+  "com.affine.payment.subscription-status.past-due": "Past-due bill",
+  "com.affine.payment.subscription-status.trialing": "Trialing",
   "com.affine.payment.benefit-1": "Unlimited local workspaces",
   "com.affine.payment.benefit-2": "Unlimited login devices",
   "com.affine.payment.benefit-3": "Unlimited blocks",
